refactor: lotto 의 사이즈, 중복값 검사를 token 단계에서 Lotto 클래스 생성 단계로 이동
